requirements.txt
Flask
Flask-WTF
Flask-Assets
Flask-Script
Flask-Mail
jsmin
cssmin
sass
Jinja2
SQLAlchemy
alembic
psycopg2
PyYAML
Pygments
passlib
pytz
py-bcrypt
requests

